requirements.txt
appdirs==1.4.3
certifi==2017.7.27.1
chardet==3.0.4
Django==1.10.8
django-braces==1.11.0
django-countries==4.6.1
django-filter==1.0.4
django-guardian==1.4.9
django-merged-inlines==1.0.1
django-oauth-toolkit==1.0.0
django-selectable==1.0.0
djangorestframework==3.6.4
fuzzywuzzy==0.15.0
idna==2.6
Markdown==2.6.8
mod_wsgi==4.5.17
oauthlib==2.0.3
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.1
python-Levenshtein==0.12.0
pytz==2017.2
requests==2.18.4
six==1.10.0
Unidecode==0.4.21
urllib3==1.22